removed leaderboard/winners views
