+++ components/brain/right-sidebar/BrainRightSidebar.tsx
@@ -10,25 +10,16 @@ import { WaveContent } from "./WaveContent";
 import { QueryKey } from "@/components/react-query-wrapper/ReactQueryWrapper";
 import { useSidebarState } from "../../../hooks/useSidebarState";
 import { ChevronDoubleRightIcon } from "@heroicons/react/24/outline";
+import { Mode } from "./BrainRightSidebarTypes";
+import type { SidebarTab } from "./BrainRightSidebarTypes";
+
+export { Mode, SidebarTab } from "./BrainRightSidebarTypes";
 
 interface BrainRightSidebarProps {
   readonly waveId: string | null | undefined;
   readonly activeTab: SidebarTab;
   readonly setActiveTab: (tab: SidebarTab) => void;
   readonly variant?: "overlay" | "inline" | undefined;
-}
-
-export enum Mode {
-  CONTENT = "CONTENT",
-  FOLLOWERS = "FOLLOWERS",
-}
-
-export enum SidebarTab {
-  ABOUT = "ABOUT",
-  LEADERBOARD = "LEADERBOARD",
-  WINNERS = "WINNERS",
-  TOP_VOTERS = "TOP_VOTERS",
-  ACTIVITY_LOG = "ACTIVITY_LOG",
 }
 
 interface SidebarState {
@@ -45,10 +36,14 @@ const useBrainRightSidebarState = (
 
   const { data: wave } = useQuery<ApiWave>({
     queryKey: [QueryKey.WAVE, { wave_id: waveId }],
-    queryFn: async () =>
-      await commonApiFetch<ApiWave>({
+    queryFn: async () => {
+      if (!waveId) {
+        throw new Error("Wave id is required");
+      }
+      return commonApiFetch<ApiWave>({
         endpoint: `waves/${waveId}`,
-      }),
+      });
+    },
     enabled: !!waveId,
     staleTime: 60000,
     placeholderData: keepPreviousData,
