+++ components/brain/right-sidebar/BrainRightSidebarTypes.ts
@@ -1,0 +1,10 @@
+export enum Mode {
+  CONTENT = "CONTENT",
+  FOLLOWERS = "FOLLOWERS",
+}
+
+export enum SidebarTab {
+  ABOUT = "ABOUT",
+  TOP_VOTERS = "TOP_VOTERS",
+  ACTIVITY_LOG = "ACTIVITY_LOG",
+}
